thumbnail gallery video youtube

+++ app/Http/Controllers/Admin/core/GalleryController.php
@@ -134,59 +134,69 @@ class GalleryController extends Controller
             File::makeDirectory($this->path);
         }
 
-        #MEMBUAT NAME FILE DARI GABUNGAN TIMESTAMP DAN UNIQID()
-        $fileName = 'Gallery' . '_' .date('Ymdhis'). '.' . $file->getClientOriginalExtension();
-
-        $size   = getimagesize($file);
-        $width  = $size[0];
-        $height = $size[1];
-
-        if($width > $height){
-            $size = ($width/$height);
-        }else{
-            $size = ($height/$width);
-        }
-
-        #UPLOAD ORIGINAN FILE (BELUM DIUBAH DIMENSINYA)
-        Image::make($file)->save($this->path . '/' . $fileName);
-        foreach ($this->dimensions as $row) {
-            #MEMBUAT CANVAS IMAGE SEBESAR DIMENSI YANG ADA DI DALAM ARRAY 
-            if($width < $height){
-                $canvas = Image::canvas($row, ceil($row*$size));
-                $resizeImage  = Image::make($file)->resize($row, ceil($row*$size), function($constraint) {
-                    $constraint->aspectRatio();
-                });
+        if ($file) {
+            #MEMBUAT NAME FILE DARI GABUNGAN TIMESTAMP DAN UNIQID()
+            $fileName = 'Gallery' . '_' .date('Ymdhis'). '.' . $file->getClientOriginalExtension();
+    
+            $size   = getimagesize($file);
+            $width  = $size[0];
+            $height = $size[1];
+    
+            if($width > $height){
+                $size = ($width/$height);
             }else{
-                $canvas = Image::canvas(($row*$size), $row);
-                $resizeImage  = Image::make($file)->resize(ceil($row*$size), $row, function($constraint) {
-                    $constraint->aspectRatio();
-                });
+                $size = ($height/$width);
             }
-
-            #CEK JIKA FOLDERNYA BELUM ADA
-            if (!File::isDirectory($this->path . '/' . $row)) {
-                #MAKA BUAT FOLDER DENGAN NAMA DIMENSI
-                File::makeDirectory($this->path . '/' . $row);
+    
+            #UPLOAD ORIGINAN FILE (BELUM DIUBAH DIMENSINYA)
+            Image::make($file)->save($this->path . '/' . $fileName);
+            foreach ($this->dimensions as $row) {
+                #MEMBUAT CANVAS IMAGE SEBESAR DIMENSI YANG ADA DI DALAM ARRAY 
+                if($width < $height){
+                    $canvas = Image::canvas($row, ceil($row*$size));
+                    $resizeImage  = Image::make($file)->resize($row, ceil($row*$size), function($constraint) {
+                        $constraint->aspectRatio();
+                    });
+                }else{
+                    $canvas = Image::canvas(($row*$size), $row);
+                    $resizeImage  = Image::make($file)->resize(ceil($row*$size), $row, function($constraint) {
+                        $constraint->aspectRatio();
+                    });
+                }
+    
+                #CEK JIKA FOLDERNYA BELUM ADA
+                if (!File::isDirectory($this->path . '/' . $row)) {
+                    #MAKA BUAT FOLDER DENGAN NAMA DIMENSI
+                    File::makeDirectory($this->path . '/' . $row);
+                }
+    
+                #MEMASUKAN IMAGE YANG TELAH DIRESIZE KE DALAM CANVAS
+                $canvas->insert($resizeImage, 'center');
+                #SIMPAN IMAGE KE DALAM MASING-MASING FOLDER (DIMENSI)
+              
+                $canvas->save($this->path . '/500/' . $fileName);
             }
-
-            #MEMASUKAN IMAGE YANG TELAH DIRESIZE KE DALAM CANVAS
-            $canvas->insert($resizeImage, 'center');
-            #SIMPAN IMAGE KE DALAM MASING-MASING FOLDER (DIMENSI)
-          
-            $canvas->save($this->path . '/500/' . $fileName);
         }
 
         $data = [
             'id_category'   => $request->kategori_produk,
             'is_created'    => \Session::get('id'),
             'embed'         => $embed,
-            'image'         => $fileName
         ];
+
+        if ($file && $embed == NULL)
+            $data['image'] = $fileName;
+        else if (!$file && $embed != NULL)
+            $data['image'] = NULL;
+
+        // dd($data);
 
         $gallery = Gallery::findOrFail($id);
 
-        File::delete($this->path.'/'.$gallery->image);
-        File::delete($this->path.'/500/'.$gallery->image);
+        if ($file || $embed != NULL) {
+            File::delete($this->path.'/'.$gallery->image);
+            File::delete($this->path.'/500/'.$gallery->image);
+        }
 
         $gallery->update($data);
